+++ frontend/src/AccountScreen/AccountScreen.js
@@ -1,0 +1,7 @@
+function AccountPage() {
+    return (
+        <p>AccountPage</p>
+    );
+}
+
+export default AccountPage;
