+++ frontend/src/HomeScreen/HomeScreen.js
@@ -1,0 +1,7 @@
+function HomeScreen() {
+    return (
+        <p>HomeScreen</p>
+    );
+}
+
+export default HomeScreen;
